Update units tests

+++ rugs_app/tests/test_orders.py
@@ -25,14 +25,22 @@ class RugsViewTest(TestCase):
         cls.orders[0].rugs.add(cls.rugs[1])
         cls.orders[1].rugs.add(cls.rugs[3])
 
+    def login_as_user(self, username, password):
+        response = self.client.post(
+            "/api/login",
+            {"username": username, "password": password},
+            "application/json"
+        )
+        return response.json()["token"]
+
     def test_get_orders_not_authenticated(self):
         orders_response = self.client.get("/api/order")
-        self.assertEqual(orders_response.status_code, 403)
+        self.assertEqual(orders_response.status_code, 401)
 
     def test_get_orders_authenticated(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
-        orders_response = self.client.get("/api/order")
+        orders_response = self.client.get("/api/order", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
         self.assertEqual(orders_response.json()["count"], 1)
         self.assertEqual(orders_response.json()["results"][0]["user"], self.regular_user1.pk)
@@ -41,9 +49,9 @@ class RugsViewTest(TestCase):
         self.assertEqual(orders_response.json()["results"][0]["status"], Order.OrderStatus.PENDING)
 
     def test_get_orders_admin(self):
-        self.client.login(username=self.superuser.username, password="admin")
+        token = self.login_as_user(username=self.superuser.username, password="admin")
 
-        orders_response = self.client.get("/api/order")
+        orders_response = self.client.get("/api/order", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
         self.assertEqual(orders_response.json()["count"], 2)
 
@@ -63,37 +71,38 @@ class RugsViewTest(TestCase):
             {"rug": self.rugs[0].pk},
             "application/json"
         )
-        self.assertEqual(add_to_cart_response.status_code, 403)
+        self.assertEqual(add_to_cart_response.status_code, 401)
 
         place_order_response = self.client.post("/api/order")
-        self.assertEqual(place_order_response.status_code, 403)
+        self.assertEqual(place_order_response.status_code, 401)
 
     def test_place_order_authenticated(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
         add_to_cart_response = self.client.post(
             "/api/cart",
             {"rug": self.rugs[0].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(add_to_cart_response.status_code, 201)
 
-        existing_rug_in_cart_response = self.client.get(f"/api/cart/{self.rugs[0].pk}")
+        existing_rug_in_cart_response = self.client.get(f"/api/cart/{self.rugs[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(existing_rug_in_cart_response.status_code, 200)
         self.assertEqual(existing_rug_in_cart_response.json()["id"], self.rugs[0].pk)
 
-        nonexistent_rug_in_cart_response = self.client.get(f"/api/cart/{self.rugs[1].pk}")
+        nonexistent_rug_in_cart_response = self.client.get(f"/api/cart/{self.rugs[1].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(nonexistent_rug_in_cart_response.status_code, 404)
 
-        place_order_response = self.client.post("/api/order")
+        place_order_response = self.client.post("/api/order", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(place_order_response.status_code, 201)
 
-        cart_response = self.client.get("/api/cart")
+        cart_response = self.client.get("/api/cart", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(cart_response.status_code, 200)
         self.assertEqual(len(cart_response.json()["cart"]), 0)
         self.assertEqual(cart_response.json()["price"], 0)
 
-        orders_response = self.client.get("/api/order")
+        orders_response = self.client.get("/api/order", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
         self.assertEqual(orders_response.json()["count"], 2)
         self.assertEqual(len(orders_response.json()["results"][1]["rugs"]), 1)
@@ -102,74 +111,80 @@ class RugsViewTest(TestCase):
         self.assertEqual(ordered_rug.status, Rug.RugStatus.NOT_AVAILABLE)
 
     def test_get_cart(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
         add_to_cart_response = self.client.post(
             "/api/cart",
             {"rug": self.rugs[0].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(add_to_cart_response.status_code, 201)
 
-        cart_response = self.client.get("/api/cart")
+        cart_response = self.client.get("/api/cart", HTTP_AUTHORIZATION=f"Token {token}")
 
         self.assertEqual(cart_response.status_code, 200)
         self.assertEqual(len(cart_response.json()["cart"]), 1)
         self.assertEqual(float(cart_response.json()["price"]), self.rugs[0].price)
 
     def test_delete_from_cart(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
         self.client.post(
             "/api/cart",
             {"rug": self.rugs[0].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.client.post(
             "/api/cart",
             {"rug": self.rugs[2].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
 
-        delete_from_cart_response = self.client.delete(f"/api/cart/{self.rugs[0].pk}")
+        delete_from_cart_response = self.client.delete(f"/api/cart/{self.rugs[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(delete_from_cart_response.status_code, 204)
 
-        cart_response = self.client.get("/api/cart")
+        cart_response = self.client.get("/api/cart", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(cart_response.status_code, 200)
         self.assertEqual(len(cart_response.json()["cart"]), 1)
         self.assertEqual(float(cart_response.json()["price"]), self.rugs[2].price)
 
     def test_delete_entire_cart(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
         self.client.post(
             "/api/cart",
             {"rug": self.rugs[0].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.client.post(
             "/api/cart",
             {"rug": self.rugs[2].pk},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
 
-        delete_response = self.client.delete("/api/cart")
+        delete_response = self.client.delete("/api/cart", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(delete_response.status_code, 204)
 
-        cart_response = self.client.get("/api/cart")
+        cart_response = self.client.get("/api/cart", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(cart_response.status_code, 200)
         self.assertEqual(len(cart_response.json()["cart"]), 0)
         self.assertEqual(float(cart_response.json()["price"]), 0)
 
     def test_place_order_no_rugs(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
         place_order_response = self.client.post(
-            "/api/order"
+            "/api/order",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(place_order_response.status_code, 400)
 
-        orders_response = self.client.get("/api/order")
+        orders_response = self.client.get("/api/order", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
         self.assertEqual(orders_response.json()["count"], 1)
 
@@ -178,15 +193,15 @@ class RugsViewTest(TestCase):
         self.assertEqual(orders_response.status_code, 403)
 
     def test_get_order_by_id_wrong_user(self):
-        self.client.login(username=self.regular_user2.username, password="test")
+        token = self.login_as_user(username=self.regular_user2.username, password="test")
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 403)
 
     def test_get_order_by_id_correct_user(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
 
         self.assertEqual(orders_response.json()["order"]["user"], self.regular_user1.pk)
@@ -194,9 +209,9 @@ class RugsViewTest(TestCase):
         self.assertEqual(orders_response.json()["order"]["status"], Order.OrderStatus.PENDING)
 
     def test_get_order_by_id_admin(self):
-        self.client.login(username=self.superuser.username, password="admin")
+        token = self.login_as_user(username=self.superuser.username, password="admin")
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
 
     def test_modify_order_not_authenticated(self):
@@ -205,42 +220,45 @@ class RugsViewTest(TestCase):
             {"status": Order.OrderStatus.COMPLETE},
             "application/json"
         )
-        self.assertEqual(modify_order_response.status_code, 403)
+        self.assertEqual(modify_order_response.status_code, 401)
 
     def test_modify_order_wrong_user(self):
-        self.client.login(username=self.regular_user2.username, password="test")
+        token = self.login_as_user(username=self.regular_user2.username, password="test")
         modify_order_response = self.client.patch(
             f"/api/order/{self.orders[0].pk}",
             {"status": Order.OrderStatus.COMPLETE},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(modify_order_response.status_code, 403)
 
     def test_modify_order_correct_user(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
         modify_order_response = self.client.patch(
             f"/api/order/{self.orders[0].pk}",
             {"status": Order.OrderStatus.COMPLETE},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(modify_order_response.status_code, 200)
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
 
         self.assertEqual(orders_response.json()["order"]["status"], Order.OrderStatus.COMPLETE)
         self.assertIsNotNone(orders_response.json()["order"]["date_completed"])
 
     def test_modify_order_admin(self):
-        self.client.login(username=self.superuser.username, password="admin")
+        token = self.login_as_user(username=self.superuser.username, password="admin")
         modify_order_response = self.client.patch(
             f"/api/order/{self.orders[0].pk}",
             {"status": Order.OrderStatus.READY_FOR_PICKUP},
-            "application/json"
+            "application/json",
+            HTTP_AUTHORIZATION=f"Token {token}"
         )
         self.assertEqual(modify_order_response.status_code, 200)
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 200)
 
         self.assertEqual(orders_response.json()["order"]["status"], Order.OrderStatus.READY_FOR_PICKUP)
@@ -248,27 +266,27 @@ class RugsViewTest(TestCase):
 
     def test_delete_order_not_authenticated(self):
         delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}")
-        self.assertEqual(delete_order_response.status_code, 403)
+        self.assertEqual(delete_order_response.status_code, 401)
 
     def test_delete_order_wrong_user(self):
-        self.client.login(username=self.regular_user2.username, password="test")
-        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}")
+        token = self.login_as_user(username=self.regular_user2.username, password="test")
+        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(delete_order_response.status_code, 403)
 
     def test_delete_order_correct_user(self):
-        self.client.login(username=self.regular_user1.username, password="test")
+        token = self.login_as_user(username=self.regular_user1.username, password="test")
 
-        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}")
+        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(delete_order_response.status_code, 204)
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 404)
 
     def test_delete_order_admin(self):
-        self.client.login(username=self.superuser.username, password="admin")
+        token = self.login_as_user(username=self.superuser.username, password="admin")
 
-        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}")
+        delete_order_response = self.client.delete(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(delete_order_response.status_code, 204)
 
-        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}")
+        orders_response = self.client.get(f"/api/order/{self.orders[0].pk}", HTTP_AUTHORIZATION=f"Token {token}")
         self.assertEqual(orders_response.status_code, 404)
